rearrange story
