+++ src/interface/src/styleguide/modal/modal.component.html
@@ -8,7 +8,7 @@
     <button>
       <mat-icon>help_outline</mat-icon>
     </button>
-    <button mat-icon-button (click)="cancel()" aria-label="close">
+    <button mat-icon-button (click)="handleCancel()" aria-label="close">
       <mat-icon class="close-btn">close</mat-icon>
     </button>
   </div>
